hugetlbfs: add O_TMPFILE support
O_TMPFILE is an option used to create an unnamed temporary regular file. 
Currently, libhugetlbfs and Seastar use a combination of mkstemp and
unlink to accomplish similar functionality on hugetlbfs.  Add O_TMPFILE
support to hugetlbfs so that it can potentially be used by existing users
(Seastar) and new users (Oracle DB).  Support is added by simply using the
d_tmpfile utility function to instantiate the dcache entry for a temporary
file.

Tested manually by successfully creating a temporary file by opening it
with (O_TMPFILE|O_RDWR) on mounted hugetlbfs and successfully mapping 2M
huge pages with it.  Without the patch, trying to open a file with
O_TMPFILE results in -ENOSUP.

[mike.kravetz@oracle.com: changelog updates]
Link: http://lkml.kernel.org/r/22c29acf9c51dae17802e1b05c9e5e4051448c5c.1571129593.git.p.sarna@tlen.pl
Signed-off-by: Piotr Sarna <p.sarna@tlen.pl>
Reviewed-by: Mike Kravetz <mike.kravetz@oracle.com>
Cc: Michal Hocko <mhocko@kernel.org>
Cc: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>