hugetlbfs: convert macros to static inline, fix sparse warning
huge_pte_offset() produced a sparse warning due to an improper return type
when the kernel was built with !CONFIG_HUGETLB_PAGE.  Fix the bad type and
also convert all the macros in this block to static inline wrappers.  Two
existing wrappers in this block had lines in excess of 80 columns so clean
those up as well.

No functional change.

Link: http://lkml.kernel.org/r/9b689b27-0291-c8b2-95aa-d3a74eab6543@oracle.com
Signed-off-by: Mike Kravetz <mike.kravetz@oracle.com>
Reported-by: Ben Dooks <ben.dooks@codethink.co.uk>
Suggested-by: Jason Gunthorpe <jgg@ziepe.ca>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>