hugetlb: remove duplicated code
Remove duplicated code between region_chg and region_add, and refactor it
into a common function, add_reservation_in_range.  This is mostly done
because there is a follow up change in another series that disables region
coalescing in region_add, and I want to make that change in one place
only.  It should improve maintainability anyway on its own.

Link: http://lkml.kernel.org/r/20190919200428.188797-3-almasrymina@google.com
Signed-off-by: Mina Almasry <almasrymina@google.com>
Reviewed-by: Mike Kravetz <mike.kravetz@oracle.com>
Cc: David Rientjes <rientjes@google.com>
Cc: Shakeel Butt <shakeelb@google.com>
Cc: Greg Thelen <gthelen@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>